Move dot net fields to separate file
