+++ src/component/child-parent-event-handler.js
@@ -1,0 +1,9 @@
+/**
+ * Handles events back and forth between child and parent react components
+ */
+export default class ChildParentEventHandler{
+    constructor(sendData, getData){
+        this.sendData = sendData;
+        this.getData = getData;
+    }
+}
