+++ src/component/file-picker-field.jsx
@@ -5,14 +5,25 @@ import './file-picker-field.css';
 
 export default function FilePickerField(props) {
     const [modalState, SetModal] = useState(false);
-    const [file, SetFile] = useState({name:'', path:'', lastModified:0, lastModifiedDate:{}, size:0, type:''});
+    const [file, SetFile] = useState({name:'', path:'', lastModified:0, lastModifiedDate:{}, size:0, type:'', extension:''});
     const [fileSelectError, SetFileSelectError] = useState("Error: could not select file");
     const modalId = `modal${props.id}`;
     const label = (props.label == undefined ? 'Select File' : props.label);
 
     function handleFileSelected(fileObject){
-        SetFile(fileObject);
-        props.onFileSelectedEvent(fileObject);
+
+        var extension = getFileExtension(fileObject.name);
+        var selectedFile = {
+            name:fileObject.name, 
+            path:fileObject.path, 
+            lastModified:fileObject.lastModified, 
+            lastModifiedDate:fileObject.lastModifiedDate, 
+            size:fileObject.size, 
+            type:fileObject.type, 
+            extension:extension
+        }
+        SetFile(selectedFile);
+        props.onFileSelectedEvent(selectedFile);
     }
 
     function handleFileError(fileObject){
@@ -24,12 +35,17 @@ export default function FilePickerField(props) {
         SetModal(false);
     }
 
+    function getFileExtension(filePath) {
+        const match = filePath.match(/\.([a-zA-Z0-9]+)$/);
+        return match ? match[1] : '';
+    }
+
     return <div>
         <Modal showModal={modalState} modalClosedEvent={hideModal} id={modalId}><p>{fileSelectError}</p></Modal>
 
         <div className='file-picker'>
             <label className="file-picker-label">{file.path}</label>
-            <FilePicker extensions={[props.fileExtension]} onChange={FileObject => handleFileSelected(FileObject)} onError={errMsg => handleFileError(errMsg)}>
+            <FilePicker extensions={props.fileExtension} onChange={FileObject => handleFileSelected(FileObject)} onError={errMsg => handleFileError(errMsg)}>
                 <button className="file-picker-button">
                     {label}
                 </button>
