+++ src/component/handleChange.js
@@ -1,9 +1,21 @@
-export default function handleChange(event, setter) {
+export default function handleChange(event, setter, postProcessing) {
+    
     setter(event.target.value);
+
+    if(postProcessing != undefined){
+        postProcessing();
+    }
+
     return event.target.value;
 }
 
-export function handleCheckboxChange(event, setter) {
+export function handleCheckboxChange(event, setter, postProcessing) {
+
     setter(event.target.checked);
+
+    if(postProcessing != undefined){
+        postProcessing();
+    }
+
     return event.target.checked;
 }
